extracted zoom from NavmapView
