+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/NavMap/NavmapZoom.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bb7799ed4671e6748979ae8dbe3553a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
